refactor: extract shared CLI utilities to reduce duplication
- Create scripts/lib/cli-utils.mjs with shared helpers
- slugify, today, findUniqueSlug, escapeYamlString, collectSlugs
- TAG_CATEGORIES shared between both CLI tools
- Reduces maintenance burden and ensures consistency
